2026-05-10 增加简单pinctrl子系统；新增pipe系统调用；新增简单信号处理支持，目前仅支持kill；vfs支持相对路径解析；

+++ include/os/types.h
@@ -90,7 +90,7 @@ typedef enum { false, true } bool;
 #define __user
 #define __iomem
 
-typedef u32 dev_t;
+typedef unsigned long dev_t;
 
 #define MINORBITS	20
 #define MINORMASK	((1U << MINORBITS) - 1)
